Rename Account Menu files & package
# Conflicts:
#	app/src/main/java/com/infomaniak/mail/ui/main/menu/user/SwitchUserAccountsAdapter.kt
#	app/src/main/java/com/infomaniak/mail/ui/main/menu/user/SwitchUserFragment.kt
#	app/src/main/java/com/infomaniak/mail/ui/main/menu/user/SwitchUserMailboxesAdapter.kt
