+++ app/src/main/res/layout/item_switch_user_account.xml
@@ -103,7 +103,7 @@
         android:visibility="gone"
         app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
         tools:itemCount="3"
-        tools:listitem="@layout/item_setting_address"
+        tools:listitem="@layout/item_switch_user_mailbox"
         tools:visibility="visible" />
 
 </LinearLayout>
