+++ app/src/main/res/layout/item_switch_user_mailbox.xml
